% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_censorship.R
\name{add_uniform_noise}
\alias{add_uniform_noise}
\title{Add uniform noise}
\usage{
add_uniform_noise(value, n_rep, sd, use_zero = FALSE)
}
\arguments{
\item{value}{a single or vector of numeric values}

\item{n_rep}{the number of replicates to make (numeric). Default is 1.}

\item{sd}{the standard deviation of the data}

\item{use_zero}{logical, should returned values be around zero or not?}
}
\value{
numeric matrix
}
\description{
Adds uniform noise to values, generating replicates with noise added to
the original.
}
