% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_censorship.R
\name{calculate_matrix_medians}
\alias{calculate_matrix_medians}
\title{Calculate matrix medians}
\usage{
calculate_matrix_medians(in_matrix, use = "col", ...)
}
\arguments{
\item{in_matrix}{numeric matrix of values}

\item{use}{character of "col" or "row" defining columns or rows}

\item{...}{extra parameters to the median function}
}
\value{
numeric
}
\description{
Given a matrix of data, calculates the median value in each column or row.
}
