% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_correlations.R
\name{cor_fast}
\alias{cor_fast}
\title{Fast correlation with test}
\usage{
cor_fast(
  x,
  y = NULL,
  use = "everything",
  method = "pearson",
  alternative = "two.sided",
  continuity = FALSE,
  return_matrix = TRUE
)
}
\arguments{
\item{x}{a numeric vector, matrix, or data frame.}

\item{y}{NULL (default) or a vector.}

\item{use}{an optional character string giving a method for computing correlations in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", or "pairwise.complete.obs".}

\item{method}{which correlation method to use, "pearson" or "spearman"}

\item{alternative}{how to perform the statistical test}

\item{continuity}{should a continuity correction be applied}

\item{return_matrix}{should the matrices of values be returned, or a long data.frame}
}
\value{
a list of matrices, rho, pvalue, or a data.frame.
}
\description{
Allows to run \code{cor.test} on a matrix of inputs.
}
\details{
Although the interface is \emph{mostly} identical to the built-in
\code{\link[stats:cor.test]{stats::cor.test()}} method, there are some differences.
\itemize{
\item if only \code{x} is provided as a matrix, the columns must be named.
\item if providing both \code{x} and \code{y}, it is assumed they are both
single vectors.
\item if \code{NA} values are present, this function does not error, but will either remove them
or return \code{NA}, depending on the option.
\item "na.or.complete" is not a valid option for \code{use}.
\item A named list with matrices or data.frame is returned, with the \code{rho} and \code{pvalue} values.
}
}
