% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{cor_matrix_2_long_df}
\alias{cor_matrix_2_long_df}
\title{convert matrix to data.frame}
\usage{
cor_matrix_2_long_df(in_matrix)
}
\arguments{
\item{in_matrix}{the correlation matrix}
}
\value{
data.frame
}
\description{
Given a square correlation matrix, converts it to a long data.frame, with three columns.
}
\details{
The data.frame contains three columns:
\itemize{
\item s1: the first entry of comparison
\item s2: the second entry of comparison
\item cor: the correlation value
}
}
