% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{disable_logging}
\alias{disable_logging}
\title{turn logging off}
\usage{
disable_logging()
}
\description{
There may be good reasons to turn the logging off after it's been turned on. This
basically tells the package that the logger isn't available.
}
