% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{enable_logging}
\alias{enable_logging}
\title{turn logging on}
\usage{
enable_logging(log_file = NULL, memory = FALSE)
}
\arguments{
\item{log_file}{the file to log to}

\item{memory}{provide memory logging too? Only available on Linux and MacOS}
}
\description{
Choose to enable logging, to a specific file if desired.
}
\details{
Uses the {logger} package under the hood, which is suggested in the dependencies.
Having logging enabled is nice to see when things are starting and stopping, and what exactly
has been done, without needing to write messages to the console. It is especially
useful if you are getting errors, but can't really see them, then you can add
"memory" logging to see if you are running out of memory.

Default log file has the pattern:

YYYY.MM.DD.HH.MM.SS_ScanCentricPeakCharacterization_run.log
}
