% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendalltau.R
\name{ici_kendalltau}
\alias{ici_kendalltau}
\title{information-content-informed kendall tau}
\usage{
ici_kendalltau(
  data_matrix,
  global_na = c(NA, Inf, 0),
  perspective = "global",
  scale_max = TRUE,
  diag_good = TRUE,
  check_timing = FALSE
)
}
\arguments{
\item{data_matrix}{samples are rows, features are columns}

\item{global_na}{globally, what should be treated as NA?}

\item{perspective}{how to treat missing data in denominator and ties, see details}

\item{scale_max}{should everything be scaled compared to the maximum correlation?}

\item{diag_good}{should the diagonal entries reflect how many entries in the sample were "good"?}

\item{check_timing}{should we try to estimate run time for full dataset? (default is FALSE)}
}
\value{
numeric
}
\description{
Given a data-matrix, computes the information-theoretic Kendall-tau-b between
all samples.
}
\details{
For more details, see the ICI-Kendall-tau vignette

\code{browseVignettes("ICIKendallTau")}

The default for \code{global_na} includes what values in the data to replace with NA for the Kendall-tau calculation. By default these are \code{global_na = c(NA, Inf, 0)}. If you want to replace something other than 0, for example, you might use \code{global_na = c(NA, Inf, -2)}, and all values of -2 will be replaced instead of 0.

When \code{check_timing = TRUE}, 5 random pairwise comparisons will be run to generate timings on a single core, and then estimates of how long the full set will take are calculated. The data is returned as a data.frame, and will be on the low side, but it should provide you with a good idea of how long your data will take.

Eventually, we plan to provide two more parameters for replacing values, \code{feature_na} for feature specific NA values and \code{sample_na} for sample specific NA values.
}
