% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendalltau.R
\name{ici_kendalltau_ref}
\alias{ici_kendalltau_ref}
\title{information-content-informed kendall tau}
\usage{
ici_kendalltau_ref(
  data_matrix,
  exclude_na = TRUE,
  exclude_inf = TRUE,
  exclude_0 = TRUE,
  zero_value = 0,
  perspective = "global",
  scale_max = TRUE,
  diag_good = TRUE,
  progress = FALSE
)
}
\arguments{
\item{data_matrix}{samples are rows, features are columns}

\item{exclude_na}{should NA values be treated as NA?}

\item{exclude_inf}{should Inf values be treated as NA?}

\item{exclude_0}{should zero values be treated as NA?}

\item{zero_value}{what is the actual zero value?}

\item{perspective}{how to treat missing data in denominator and ties, see details}

\item{scale_max}{should everything be scaled compared to the maximum correlation?}

\item{diag_not_na}{should the diagonal entries reflect how many entries in the sample were "good"?}
}
\value{
numeric
}
\description{
Given a data-matrix, computes the information-content-informed (ICI) Kendall-tau-b between
all samples.
}
\details{
For more details, see the ICI-Kendall-tau vignette:
\href{../doc/ici-kendalltau.html}{\code{vignette("ici-kendalltau", package = "ICIKendallTau")}}
}
\keyword{internal}
