% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ici_kt}
\alias{ici_kt}
\title{Calculates ici-kendall-tau}
\usage{
ici_kt(
  x,
  y,
  perspective = "local",
  alternative = "two.sided",
  continuity = FALSE,
  output = "simple"
)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{perspective}{should we consider the "local" or "global" perspective?}

\item{alternative}{what is the alternative for the p-value test?}

\item{continuity}{logical: if true, a continuity correction is used}

\item{output}{used to control reporting of values for debugging}
}
\value{
kendall tau correlation
}
\description{
Calculates ici-kendall-tau
}
\details{
Calculates the information-content-informed Kendall-tau correlation measure.
This correlation is based on concordant and discordant ranked pairs, like Kendall-tau,
but also includes missing values (as NA). Missing values are assumed to be \emph{primarily} due
to lack of detection due to instrumental sensitivity, and therefore encode \emph{some} information.

For more details see the ICI-Kendall-tau vignette:

\code{browseVignettes("ICIKendallTau")}
}
\examples{
x = sort(rnorm(100))
y = x + 1
y2 = y
y2[1:10] = NA
ici_kt(x, y)
ici_kt(x, y2, "global")
ici_kt(x, y2)

}
