% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendalltau.R
\name{kt_fast}
\alias{kt_fast}
\title{fast kendall tau}
\usage{
kt_fast(x, y = NULL, use = "everything", return_matrix = TRUE)
}
\arguments{
\item{x}{a numeric vector, matrix, or data frame.}

\item{y}{NULL (default) or a vector.}

\item{use}{an optional character string giving a method for computing correlations in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", or "pairwise.complete.obs".}

\item{return_matrix}{Should the matrices of values be returned, or a long data.frame}
}
\value{
a named vector or list of matrices.
}
\description{
Uses the underlying c++ implementation of \code{ici_kt} to provide a fast version
of Kendall-tau correlation.
}
\details{
Although the interface is \emph{mostly} identical to the built-in \code{stats::cor} method,
there are some differences.
\itemize{
\item if providing both \code{x} and \code{y}, it is assumed they are both
single vectors.
\item if \code{NA} values are present, this function does not error, but will either remove them
or return \code{NA}, depending on the option.
\item "na.or.complete" is not a valid option for \code{use}.
\item A named vector or a named list with matrices is returned, with the \code{tau} and \code{pvalue} values.
}
}
