% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_memory}
\alias{log_memory}
\title{log memory usage}
\usage{
log_memory()
}
\description{
Logs the amount of memory being used to a log file if it is available, and
generating warnings if the amount of RAM hits zero.
}
