% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_message}
\alias{log_message}
\title{log messages}
\usage{
log_message(message_string)
}
\arguments{
\item{message_string}{the string to put in the message}
}
\description{
If a log_appender is available, logs the given message at the \code{info} level.
}
