% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_censorship.R
\docType{data}
\name{missing_dataset}
\alias{missing_dataset}
\title{Example Dataset With Missingness}
\format{
\subsection{\code{missing_dataset}}{

A matrix with 1000 rows and 20 columns, where rows are features and
columns are samples.
}
}
\source{
Robert M Flight
}
\usage{
missing_dataset
}
\description{
An example dataset that has missingness from left-censorship
}
\keyword{datasets}
