% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendalltau.R
\name{pairwise_completeness}
\alias{pairwise_completeness}
\title{pairwise completeness}
\usage{
pairwise_completeness(
  data_matrix,
  global_na = c(NA, Inf, 0),
  include_only = NULL,
  return_matrix = TRUE
)
}
\arguments{
\item{data_matrix}{samples are columns, features are rows}

\item{global_na}{globally, what should be treated as NA?}

\item{include_only}{is there certain comparisons to do?}

\item{return_matrix}{should the matrix or data.frame be returned?}
}
\value{
matrix of degree of completeness
}
\description{
Calculates the completeness between any two samples using "or", is an
entry missing in either X "or" Y.
}
