% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendalltau.R
\name{pairwise_missingness}
\alias{pairwise_missingness}
\title{pairwise missingness}
\usage{
pairwise_missingness(
  data_matrix,
  exclude_na = TRUE,
  exclude_inf = TRUE,
  exclude_0 = TRUE,
  zero_value = 0
)
}
\arguments{
\item{data_matrix}{samples are rows, features are columns}

\item{exclude_na}{should NA values be treated as NA?}

\item{exclude_inf}{should Inf values be treated as NA?}

\item{exclude_0}{should zero values be treated as NA?}

\item{zero_value}{what is the actual zero value?}
}
\value{
matrix of missing values
}
\description{
Calculates the missingness between any two samples using "or", is an
entry missing in either X "or" Y.
}
