% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank-ordering.R
\name{rank_order_data}
\alias{rank_order_data}
\title{Rank order row data}
\usage{
rank_order_data(data_matrix, global_na = c(NA, Inf, 0))
}
\arguments{
\item{data_matrix}{matrix or data.frame of values}

\item{global_na}{the values to consider as missing}
}
\value{
list with two matrices and a data.frame
}
\description{
Given a data-matrix of numeric data, calculates the rank of each row in each
column (feature in sample), gets the median rank across all columns, and
returns the original data with missing values set to NA, the reordered data,
and a data.frame of the ranks of each feature and the number of missing values.
}
