% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendalltau.R
\name{ref_diff_kendallt}
\alias{ref_diff_kendallt}
\title{compute kendall tau}
\usage{
ref_diff_kendallt(
  x,
  y,
  perspective = "local",
  alternative = "two.sided",
  output = "simple"
)
}
\arguments{
\item{x}{vector of x data}

\item{y}{vector of y data}

\item{perspective}{how to treat missing data, see details}
}
\value{
numeric
}
\description{
Reference version for ICI-kendall-tau. Given two vectors of data, computes the Kendall Tau correlation between them.
This version has logic for handling missing data in X and Y.
}
\examples{
data("grp_cor_data")
exp_data = grp_cor_data$data
x = exp_data[, 1]
y = exp_data[, 2]
kendallt(x, y)
cor(x, y, method = "kendall") 

x = sort(rnorm(100))
y = x + 1
y2 = y
y2[1:10] = NA
kendallt(x, y)
kendallt(x, y2, "global")
kendallt(x, y2)
}
\keyword{internal}
