% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{show_progress}
\alias{show_progress}
\title{turn progress on off}
\usage{
show_progress(progress = TRUE)
}
\arguments{
\item{progress}{logical to have it on or off}
}
\description{
Allow the user to turn progress messages to the console and off. Default
is to provide messages to the console.
}
