% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_censorship.R
\name{test_left_censorship}
\alias{test_left_censorship}
\title{Test for left censorship}
\usage{
test_left_censorship(
  data_matrix,
  global_na = c(NA, Inf, 0),
  sample_classes = NULL
)
}
\arguments{
\item{data_matrix}{matrix or data.frame of numeric data}

\item{global_na}{what represents zero or missing?}

\item{sample_classes}{which samples are in which class}
}
\value{
data.frame of trials / successes, and binom.test result
}
\description{
Does a binomial test to check if the most likely cause of missing values
is due to values being below the limit of detection, or coming from a
left-censored distribution.
}
\details{
For each feature that is missing in a group of samples, we save as a possibility
to test. For each sample, we calculate the median value with any missing values
removed. Each feature that had a missing value, we test whether the remaining
non-missing values are below the sample median for those samples where the
feature is non-missing. A binomial test considers the total number of features
instances (minus missing values) as the number of trials, and the number of
of features below the sample medians as the number of successes.

There is a bit more detail in the vignette: \code{vignette("testing-for-left-censorship", package = "ICIKendallTau")}
}
\examples{
# this example has 80\% missing due to left-censorship
data(missing_dataset)
missingness = test_left_censorship(missing_dataset)
missingness$values
missingness$binomial_test

}
\seealso{
\code{\link[=ici_kendalltau]{ici_kendalltau()}}
}
