% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_censorship.R
\docType{data}
\name{yeast_missing}
\alias{yeast_missing}
\title{Example RNA-Seq Dataset With Missingness}
\format{
\subsection{\code{yeast_missing}}{

A matrix with 6887 rows (genes) and 96 columns (samples).
}
}
\source{
\url{https://dx.doi.org/10.6084/M9.FIGSHARE.1425502.V1} \url{https://dx.doi.org/10.6084/M9.FIGSHARE.1425503.V1}
}
\usage{
yeast_missing
}
\description{
An example dataset from RNA-seq experiment on yeast, created by Gierliński et al.,
"Statistical models for RNA-seq data derived from a two-condition 48-replicate experiment",
Bioinformatics, 31, 2015 \url{https://doi.org/10.1093/bioinformatics/btv425}.
}
\keyword{datasets}
