% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendalltau.R
\name{ici_kendalltau}
\alias{ici_kendalltau}
\title{information-content-informed kendall tau}
\usage{
ici_kendalltau(
  data_matrix,
  exclude_na = TRUE,
  exclude_inf = TRUE,
  exclude_0 = TRUE,
  zero_value = 0,
  perspective = "global",
  scale_max = TRUE,
  diag_good = TRUE,
  check_timing = FALSE
)
}
\arguments{
\item{data_matrix}{samples are rows, features are columns}

\item{exclude_na}{should NA values be treated as NA?}

\item{exclude_inf}{should Inf values be treated as NA?}

\item{exclude_0}{should zero values be treated as NA?}

\item{zero_value}{what is the actual zero value?}

\item{perspective}{how to treat missing data in denominator and ties, see details}

\item{scale_max}{should everything be scaled compared to the maximum correlation?}

\item{diag_good}{should the diagonal entries reflect how many entries in the sample were "good"?}

\item{check_timing}{should we try to estimate run time for full dataset? (default is FALSE)}
}
\value{
numeric
}
\description{
Given a data-matrix, computes the information-theoretic Kendall-tau-b between
all samples.
}
\details{
For more details, see the ICI-Kendall-tau vignette

\code{browseVignettes("ICIKendallTau")}

When \code{check_timing = TRUE}, 5 random pairwise comparisons will be run to generate timings on a single core, and then estimates of how long the full set will take are calculated. The data is returned as a data.frame, and will be on the low side, but it should provide you with a good idea of how long your data will take.
}
