% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{long_df_2_cor_matrix}
\alias{long_df_2_cor_matrix}
\title{convert data.frame to matrix}
\usage{
long_df_2_cor_matrix(long_df, is_square = TRUE)
}
\arguments{
\item{long_df}{the long data.frame}
}
\value{
matrix
}
\description{
Given a long data.frame, converts it to a possibly square correlation matrix
}
